Add getting started
Adds the version of getting started with magma
page taken from the squeak wiki:
	https://wiki.squeak.org/squeak/2689

I assume, that the original author of that wiki entry
was created by Chris Muller or someone else (as the
history does not list any author, only ip addresses).

And it's ok to include it, as it's open in public.